+++ pkg/util/validation/limits_test.go
@@ -11,6 +11,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v2"
+
+	"github.com/cortexproject/cortex/pkg/util/flagext"
 )
 
 // mockTenantLimits exposes per-tenant limits based on a provided map
@@ -64,6 +66,43 @@ func TestLimits_Validate(t *testing.T) {
 
 		t.Run(testName, func(t *testing.T) {
 			assert.Equal(t, testData.expected, testData.limits.Validate(testData.shardByAllLabels))
+		})
+	}
+}
+
+func TestOverrides_MaxChunksPerQueryFromStore(t *testing.T) {
+	tests := map[string]struct {
+		setup    func(limits *Limits)
+		expected int
+	}{
+		"should return the default legacy setting with the default config": {
+			setup:    func(limits *Limits) {},
+			expected: 2000000,
+		},
+		"the new config option should take precedence over the deprecated one": {
+			setup: func(limits *Limits) {
+				limits.MaxChunksPerQueryFromStore = 10
+				limits.MaxChunksPerQuery = 20
+			},
+			expected: 20,
+		},
+		"the deprecated config option should be used if the new config option is unset": {
+			setup: func(limits *Limits) {
+				limits.MaxChunksPerQueryFromStore = 10
+			},
+			expected: 10,
+		},
+	}
+
+	for testName, testData := range tests {
+		t.Run(testName, func(t *testing.T) {
+			limits := Limits{}
+			flagext.DefaultValues(&limits)
+			testData.setup(&limits)
+
+			overrides, err := NewOverrides(limits, nil)
+			require.NoError(t, err)
+			assert.Equal(t, testData.expected, overrides.MaxChunksPerQueryFromStore("test"))
 		})
 	}
 }